Extract strategy list and list item to common
